% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_q_names.R
\name{break_q_names}
\alias{break_q_names}
\title{add line break to question names}
\usage{
break_q_names(q_names, max_lines = 2)
}
\arguments{
\item{q_names}{question names - character}

\item{max_lines}{maximum number of lines. Int, defaults to 2}
}
\value{

}
\description{
add line breaks to question names - split into two lines.
Line lengths are determined by the longest question.
}
