% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freq_ability}
\alias{calc_freq_ability}
\title{Has coding ability changed frequency table}
\usage{
calc_freq_ability(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2:: functions.}
}
\value{
data.frame
}
\description{
Function creates frequency table of coding ability changes in current role
}
