% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freq_learn_code}
\alias{calc_freq_learn_code}
\title{First learnt code frequency table}
\usage{
calc_freq_learn_code(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2:: functions.}
}
\value{
data.frame
}
\description{
Function creates frequency table about where first coding was learnt
}
