% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freq_operations}
\alias{calc_freq_operations}
\title{Operations and how they are preformed}
\usage{
calc_freq_operations(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2:: functions}
}
\value{
data.frame
}
\description{
Create frequency table of operations and how they are preformed
}
