% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freq_prior_coding}
\alias{calc_freq_prior_coding}
\title{Prior coding experience frequency table}
\usage{
calc_freq_prior_coding(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2:: functions.}
}
\value{
data.frame
}
\description{
Function creates frequency table about coding experience prior to current role
}
