% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freq_reproducible_workflow}
\alias{calc_freq_reproducible_workflow}
\title{Knowledge of reproducible workflow packages frequency table}
\usage{
calc_freq_reproducible_workflow(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2:: functions.}
}
\value{
data.frame
}
\description{
Create frequency table of knowledge of reproducible workflow packages
}
