% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freq_version_control}
\alias{calc_freq_version_control}
\title{Version control platforms frequency table}
\usage{
calc_freq_version_control(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2:: functions.}
}
\value{
data.frame
}
\description{
Function creates frequency table of use of version control software
}
