% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_coding}
\alias{calc_freqs_coding}
\title{Coding usage}
\usage{
calc_freqs_coding(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data functions.}
}
\value{
data.frame
}
\description{
Function which calculates the frequencies of time spent coding
}
