% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_coding_frequency}
\alias{calc_freqs_coding_frequency}
\title{calc_freqs_coding_frequency}
\usage{
calc_freqs_coding_frequency(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}
}
\value{
data.frame
}
\description{
calc_freqs_coding_frequency
}
