% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_coding_practice_usage}
\alias{calc_freqs_coding_practice_usage}
\title{calc_freqs_coding_practice_usage}
\usage{
calc_freqs_coding_practice_usage(data, code_prac_levels)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}

\item{code_prac_levels}{A vector of strings, length 6}
}
\value{
data.frame
}
\description{
calc_freqs_coding_practice_usage
}
