% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_coding_practices}
\alias{calc_freqs_coding_practices}
\title{Types of coding practices used}
\usage{
calc_freqs_coding_practices(data, code_prac_levels)
}
\arguments{
\item{data}{This is generated using the data functions.}

\item{code_prac_levels}{A vector of strings, length 6}
}
\value{
data.frame
}
\description{
Used for genertaing more than 1 table = calc_freqs_coding_practice_usage, calc_freqs_documenation_usage
}
