% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_coding_tool_access_knowledge}
\alias{calc_freqs_coding_tool_access_knowledge}
\title{calc_freqs_coding_tool_access_knowledge}
\usage{
calc_freqs_coding_tool_access_knowledge(data, langs)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}

\item{langs}{Formatted list of strings. langs object is available in carsurvey2::generate_tables()}
}
\value{
data.frame
}
\description{
calc_freqs_coding_tool_access_knowledge
}
