% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_knowledge_of_rap}
\alias{calc_freqs_knowledge_of_rap}
\title{Knowledge of RAP}
\usage{
calc_freqs_knowledge_of_rap(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data functions.}
}
\value{
data.frame
}
\description{
Create a frequency table of knowledge of RAP
}
