% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_opinion_of_rap}
\alias{calc_freqs_opinion_of_rap}
\title{calc_freqs_opinion_of_rap}
\usage{
calc_freqs_opinion_of_rap(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}
}
\value{
data.frame
}
\description{
calc_freqs_opinion_of_rap
}
