% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_outside_work}
\alias{calc_freqs_outside_work}
\title{Coding experience outside of work frequency table}
\usage{
calc_freqs_outside_work(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2:: functions.}
}
\value{
data.frame
}
\description{
Function creates frequency table of if any coding experience outside current role
}
