% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_practices_usage}
\alias{calc_freqs_practices_usage}
\title{Types of coding practices used}
\usage{
calc_freqs_practices_usage(data, code_prac_levels)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}

\item{code_prac_levels}{A vector of strings, length 6}
}
\value{
data.frame
}
\description{
Function creates frequency table of the types of coding practices used
}
