% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_rap_advanced}
\alias{calc_freqs_rap_advanced}
\title{RAP score for advanced components}
\usage{
calc_freqs_rap_advanced(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data functions.}
}
\value{
data.frame
}
\description{
Function creates frequency of RAP score for advanced components
}
