% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_rap_basic}
\alias{calc_freqs_rap_basic}
\title{RAP score for basic components}
\usage{
calc_freqs_rap_basic(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}
}
\value{
data.frame
}
\description{
Function creates frequency table of RAP score basic components
}
