% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_rap_score_components}
\alias{calc_freqs_rap_score_components}
\title{RAP score components}
\usage{
calc_freqs_rap_score_components(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data functions.}
}
\value{
data.frame
}
\description{
Function creates frequency table of basic and advanced RAP score components
}
