% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_multi_col_freqs.R
\name{calc_multi_col_freqs}
\alias{calc_multi_col_freqs}
\title{Calculate frequencies for multiple columns}
\usage{
calc_multi_col_freqs(cols, factor_levels, calc_props = FALSE)
}
\arguments{
\item{cols}{the selected columns (data.frame)}

\item{factor_levels}{a list of factor levels for the cell values of all columns}

\item{calc_props}{whether or not to calculate proportions (logical, FALSE by default)}
}
\value{
frequencies data frame with the frequencies of response options by column
}
\description{
Calculates frequencies for multiple dataframe columns and returns them in a single data frame which can be used with the 
plot_stacked or plot_likert functions
}
