% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_derive_code_status}
\alias{data_derive_code_status}
\title{data_derive_code_status}
\usage{
data_derive_code_status(data)
}
\arguments{
\item{data}{a data frame containing cleaned CARS wave 2 data}
}
\value{
df containing the coding tool status columns
}
\description{
Derive coding tool knowledge but no access/access but no knowledge status..
}
