% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_derive_rap_scores}
\alias{data_derive_rap_scores}
\title{data_derive_rap_scores}
\usage{
data_derive_rap_scores(data)
}
\arguments{
\item{data}{a date frame containing cleaned CARS wave 2 data}
}
\value{
df containing the additional RAP score columns
}
\description{
Calculate RAP scores. Derive RAP score columns from existing variables and add to the dataframe.
}
