% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_ingest}
\alias{data_ingest}
\title{data_ingest}
\usage{
data_ingest(
  survey = "790800",
  export = "1438876",
  token = Sys.getenv("CARS_TOKEN"),
  secret = Sys.getenv("CARS_SECRET"),
  check_hash = "6c64deda0580f9cf9cc6a32a3445472b13179053"
)
}
\arguments{
\item{survey}{the survey ID (character string/numeric). Defaults to "790800".}

\item{export}{the export ID (character string/numeric). Defaults to "1438876".}

\item{token}{the API token (character string). Loaded from environment variable by default.}

\item{secret}{the secret API token (character string). Loaded from environment variable by default.}

\item{check_hash}{a hash to check the downloaded data aginst or FALSE to skip the check (character string/logical). Defaults to final export hash.}
}
\value{
the exported data as a dataframe
}
\description{
Download smartsurvey export via the API. Download the exported data from smartsurvey using the API. Use convert_raw() to convert the API response to a data.frame.
}
