% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_rename_cols}
\alias{data_rename_cols}
\title{data_rename_cols}
\usage{
data_rename_cols(data)
}
\arguments{
\item{data}{a data frame containing prepared CARS wave 2 data}
}
\value{
df with renamed columns
}
\description{
Rename columns in CARS wave 2 dataset.
}
