% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_tidy_ingest}
\alias{data_tidy_ingest}
\title{data_tidy_ingest}
\usage{
data_tidy_ingest(data)
}
\arguments{
\item{data}{the data returned by the ingest function (data.frame)}
}
\value{
the tidied data (data.frame)
}
\description{
Tidy the ingested data from the smartsurvey API. Tidy column names and remove empty rows from the ingested data. 
The question numbers returned are only nested to one level, e.g. Q1.1, Q1.2, etc. However, the function will work on any number of nester levels or none. 
Use covert_raw() to convert the smartsurvey API response to a data.frame before using this function on the data.
}
