% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_rap_scores.R
\name{derive_rap_scores}
\alias{derive_rap_scores}
\title{Calculate RAP scores}
\usage{
derive_rap_scores(data)
}
\arguments{
\item{data}{a date frame containing cleaned CARS wave 2 data}
}
\value{
df containing the additional RAP score columns
}
\description{
Derive RAP score columns from existing variables and add to the dataframe.
}
