% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{format_department_path}
\alias{format_department_path}
\title{Format department path}
\usage{
format_department_path(department)
}
\arguments{
\item{department}{A string}
}
\value{
String
}
\description{
This function cleans the department names
}
