% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_file_path}
\alias{format_file_path}
\title{format_file_path}
\usage{
format_file_path(department)
}
\arguments{
\item{department}{}
}
\value{
String
}
\description{
This function cleans the department names
}
\examples{

department = "Cabinet Office (excl. agencies)"

format_file_path(department)
}
