% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{generate_dummy_data}
\alias{generate_dummy_data}
\title{Generate dummy data similar to carsurvey wave 2 responses}
\usage{
generate_dummy_data(dummy_data_row_number = 100, config = NULL)
}
\arguments{
\item{dummy_data_row_number}{Number of rows to generate}

\item{config}{Config used to control the data generation. Included in paackage and loaded by default.
BUT a custom config can be supplied if desired. Data generated is equal to what is produced by thecarsurvey2::data_convert_raw(API_data)  %>% carsurvey2::data_tidy_ingest() %>%  carsurvey2::data_rename_cols(). 
You can therefore run carsurvey2::data_derive_rap_scores(carsurvey2::generate_dummy_data(10))}
}
\value{
data.frame
}
\description{
Generate dummy data similar to carsurvey wave 2 responses
}
\details{
Generates non-disclosure data similar to the carsurvey wave 2 responses. 
The column names are the same, same number of column. The data is similar but the distbutions are random. Where a question was multiple choice these are sampled randomly. 
Free text is auto generated as a sting of random letters. The disrbutions are NOT the same, e.g for department this is sampled randomly so each deartment has a equal chance of being selected.
}
