% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_gradient}
\alias{get_gradient}
\title{Create single colour gradient}
\usage{
get_gradient(n, colour = c(0, 69, 86))
}
\arguments{
\item{n}{the number of colours needed}

\item{colour}{the first colour of the gradient - numeric vector with three values representing red, blue and green (max 255)}
}
\value{
RGB colours
}
\description{
Creates a list of colours for plotting. The list contains n shades of the original colour, increasing in brightness.
}
