% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_insert_table_toggle}
\alias{html_insert_table_toggle}
\title{html_insert_table_toggle}
\usage{
html_insert_table_toggle(output_name)
}
\arguments{
\item{output_name}{the name of assigned to the output. Used to name buttons and reference tables and chart.}
}
\value{
html buttons and CSS
}
\description{
Add table toggle buttons to rmarkdowndown site
}
\details{
Add "show table" and "show chart" buttons to rmarkdown site (html). Expects table contaners to be named "name-table" and chart containers to be named "name-chart"
}
