% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_likert_table.R
\name{output_likert_table}
\alias{output_likert_table}
\title{build html table forlikert outputs (kable)}
\usage{
output_likert_table(table)
}
\arguments{
\item{table}{Frequency table for likert quesitons (data frame). 4+ columns - question names in column 1 with answer options in subsequent columns. Frequencies should proportions, between 0 and 1.}
}
\value{
html table
}
\description{
build output table for likert question data
}
