% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_freqs}
\alias{plot_freqs}
\title{Plot frequency graph}
\usage{
plot_freqs(
  table,
  xlab,
  ylab,
  bar_colour,
  n,
  font_size = 12,
  orientation = "v",
  break_q_names_col = NULL,
  max_lines = 2,
  ...
)
}
\arguments{
\item{table}{Frequency table (data frame). 2 columns - cateogry names and frequencies.}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{bar_colour}{Colour name. Defaults to blue (see get_gradient())}

\item{n}{sample size}

\item{font_size}{minimum font size for the plot (numeric).}

\item{orientation}{plot orientation ("h" = horizontal, "v" = verical). Vertical by default}

\item{break_q_names_col}{applies break_q_names to the column. Not applied by default}

\item{max_lines}{maximum number of lines. Int, defaults to 2/ See carsurvey::break_q_names()}

\item{...}{additional plotly_ly arguments}
}
\value{
bar chart
}
\description{
Produce bar chart (plotly) for single factor frequency data.
}
