% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_site_yml.R
\name{read_site_yml}
\alias{read_site_yml}
\title{Read RMarkdown site yaml}
\usage{
read_site_yml(filename, text)
}
\arguments{
\item{filename}{the yaml file name (including path). Optional if text argument is present.}

\item{text}{character string - the contents of the yaml file. Optional if filename is present.}
}
\value{
list of arguments
}
\description{
Read the rmarkdown site yaml and return list of navbar arguments
}
