% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_department_pages}
\alias{render_department_pages}
\title{render_department_pages}
\usage{
render_department_pages(smart_survey_data)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::data_ functions.}
}
\description{
Creates pages for each department.
}
\details{
For each department filters the smart_survey_data to only that department
and then renders the common department template generating a unique page for each 
department. The template is located at the internal var template_path.
}
