% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_filtered_pages}
\alias{render_filtered_pages}
\title{render_department_pages}
\usage{
render_filtered_pages(
  smart_survey_data,
  filter_variable,
  page_title = "",
  output_folder = "../../docs",
  template_path = "rmarkdown/deps/template.rmd"
)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::data_ functions.}

\item{output_folder}{Folder the site is built and saved to}

\item{template_path}{The path to the template that gets render for each department}
}
\description{
Creates pages for each department.
}
\details{
For each department filters the smart_survey_data to only that department
and then renders the common department template generating a unique page for each 
department. The template is located at the internal var template_path.
}
