% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_main_site}
\alias{render_main_site}
\title{render_main_site}
\usage{
render_main_site(smart_survey_data)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::data_ functions.}
}
\description{
Processes the Smart survey data to generate a series of tables. 
Then removes the old site and uses the generated tables to render a new site.
Uses the markdown files located in the internal var markdown_file_path.
}
