% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_table_toggle.R
\name{setup_table_toggle}
\alias{setup_table_toggle}
\title{Set up JavaScript functions for show table/chart buttons}
\usage{
setup_table_toggle()
}
\value{
JavaScript code (raw html)
}
\description{
Insert JavaScript code for show table/chart buttons. Use with insert_table_toggle().
Only include once per document. Assumes the naving convention "name-table" and "name-chart" for divs containing tables and charts.
}
