% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_access_to_programming_language}
\alias{table_access_to_programming_language}
\title{table_access_to_programming_language}
\usage{
table_access_to_programming_language(smart_survey_data, langs)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}

\item{langs}{Formatted list of strings. langs object is available in carsurvey2::generate_tables()}
}
\value{
data.frame
}
\description{
table_access_to_programming_language
}
