% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_coding_frequency}
\alias{table_coding_frequency}
\title{table_coding_frequency}
\usage{
table_coding_frequency(smart_survey_data)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}
}
\value{
data.frame
}
\description{
table_coding_frequency
}
