% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_coding_tool_access_knowledge}
\alias{table_coding_tool_access_knowledge}
\title{table_coding_tool_access_knowledge}
\usage{
table_coding_tool_access_knowledge(smart_survey_data, langs)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}

\item{langs}{Formatted list of strings. langs object is available in carsurvey2::generate_tables()}
}
\value{
data.frame
}
\description{
table_coding_tool_access_knowledge
}
