% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_documenation_usage}
\alias{table_documenation_usage}
\title{table_documenation_usage}
\usage{
table_documenation_usage(smart_survey_data, code_prac_levels)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}

\item{code_prac_levels}{A vector of strings, length 6}
}
\value{
data.frame
}
\description{
table_documenation_usage
}
