% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_knowledge_of_rap}
\alias{table_knowledge_of_rap}
\title{table_knowledge_of_rap}
\usage{
table_knowledge_of_rap(smart_survey_data)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}
}
\value{
data.frame
}
\description{
table_knowledge_of_rap
}
