% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_rap_score_components}
\alias{table_rap_score_components}
\title{table_rap_score_components}
\usage{
table_rap_score_components(smart_survey_data)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}
}
\value{
data.frame
}
\description{
table_rap_score_components
}
