% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{wrap_outputs}
\alias{wrap_outputs}
\title{Wrap outputs}
\usage{
wrap_outputs(name, chart, table)
}
\arguments{
\item{name}{output name - string}

\item{chart}{plotly widget}

\item{table}{html table (string) - knitr::kable is recommended}
}
\value{
html widget containing the chart, table and toggling functionality
}
\description{
Wrap corresponding chart and table outputs and add toggle functionality
}
