% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_navbar.R
\name{build_navbar}
\alias{build_navbar}
\title{build rmarkdown site navbar}
\usage{
build_navbar(navbar_info)
}
\arguments{
\item{navbar_info}{list of arguments - use output from read_site_yml()}
}
\value{
navbar html as string
}
\description{
Build rmarkdown site navbar using arguments from the site yaml file (use the output from read_site_yml() as the navbar_info argument)
}
