% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{calc_freqs_rap_score_components}
\alias{calc_freqs_rap_score_components}
\title{calc_freqs_rap_score_components}
\usage{
calc_freqs_rap_score_components(data)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}
}
\value{
data.frame
}
\description{
calc_freqs_rap_score_components
}
