% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{coding_practices}
\alias{coding_practices}
\title{coding_practices}
\usage{
coding_practices(smart_survey_data, code_prac_levels)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}

\item{code_prac_levels}{A vector of strings, length 6}
}
\value{
data.frame
}
\description{
Used for genertaing more than 1 table = table_coding_practice_usage, table_documenation_usage
}
