% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R, R/generate_dummy_data.R
\name{generate_dummy_data}
\alias{generate_dummy_data}
\title{Generate dummy data similar to carsurvey wave 2 responses}
\usage{
generate_dummy_data(nrows = 100)

generate_dummy_data(nrows = 100)
}
\arguments{
\item{nrows}{Number of rows to generate}
}
\value{
data.frame

data.frame
}
\description{
Generate dummy data similar to carsurvey wave 2 responses

Generate dummy data similar to carsurvey wave 2 responses
}
\details{
Generates non-disclosure data similar to the carsurvey wave 2 responses. 
The column names are the same, same number of column. The data is similar but the distbutions are random.
 The disrbutions are NOT the same, e.g for department this is sampled randomly so each deartment has a equal chance of being selected.

Generates non-disclosure data similar to the carsurvey wave 2 responses. 
The column names are the same, same number of column. The data is similar but the distbutions are random.
 The disrbutions are NOT the same, e.g for department this is sampled randomly so each deartment has a equal chance of being selected.
}
