% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{generate_tables}
\alias{generate_tables}
\title{generate_tables}
\usage{
generate_tables(smart_survey_data)
}
\arguments{
\item{smart_survey_data}{This is generated using the carsurvey2::smart_survey_data_ functions.}
}
\value{

}
\description{

}
