% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_table_toggle.R
\name{insert_table_toggle}
\alias{insert_table_toggle}
\title{Add table toggle buttons to down document}
\usage{
insert_table_toggle(output_name)
}
\arguments{
\item{output_name}{the name of assigned to the output. Used to name buttons and reference tables and chart.}
}
\value{
html buttons and CSS
}
\description{
Add "show table" and "show chart" buttons to rmarkdown document (html). Expects table contaners to be named "name-table" and chart containers to be named "name-chart"
}
