% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likert.R
\name{plot_likert}
\alias{plot_likert}
\title{Plot likert graph}
\usage{
plot_likert(table, mid, xlab, ylab, n, font_size = 12, neutral_mid = TRUE, ...)
}
\arguments{
\item{table}{Frequency table for likert quesitons (data frame). 4+ columns - question names in column 1 with answer options in subsequent columns. Frequencies should proportions, between 0 and 1.}

\item{mid}{the mid-point of the scale. should be higher than 2 and lower than the number of answers.}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{n}{sample size}

\item{font_size}{minimum font size for the plot (numeric).}

\item{neutral_mid}{whether the middle of the scale should be a neutral category (logical). TRUE by default}

\item{...}{additional plot_ly arguments}
}
\value{
bar chart
}
\description{
Produce likert stacked bar chart (plotly). At least 2 questions per plot.
}
