% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_filtered_pages}
\alias{render_filtered_pages}
\title{Render filtered pages}
\usage{
render_filtered_pages(
  data,
  filter_variable,
  page_title = "",
  output_folder = "../../docs",
  template_path = "rmarkdown/summary_template/template.rmd"
)
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}

\item{filter_variable}{The variable that the data is filtered on. This is given as a string such as "dept". 
The data is filtered by getting all values in the filter_variable greater than 20 and subsetting the data.}

\item{page_title}{This is ONLY PART of the title. The title is generated from "DRAFT: ", page_title , " profile: ", filter -- 
where filter is one element of the list as described above in filter_variable (values over 20) and the page title is this argument.}

\item{output_folder}{Folder the site is built and saved to}

\item{template_path}{The path to the template that gets render for each department}
}
\description{
Creates pages by filter (e.g. by department/profession/grade).
}
\details{
For each department filters the data to only that department
and then renders the common department template generating a unique page for each 
department. The template is located at the internal var template_path.
}
