% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_navbar}
\alias{render_navbar}
\title{Render navigation bar}
\usage{
render_navbar(yml_path = "rmarkdown/main/_site.yml")
}
\arguments{
\item{yml_path}{Path to the rmarkdown site yml config}
}
\value{
navbar_page html code for navbar
}
\description{
Creates the site navbar.
}
