% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{setup_table_toggle}
\alias{setup_table_toggle}
\title{Set up table toggle}
\usage{
setup_table_toggle()
}
\value{
JavaScript code (raw html)
}
\description{
Set up JavaScript functions for show table/chart buttons
}
\details{
Insert JavaScript code for show table/chart buttons. Use with insert_table_toggle().
Only include once per document. Assumes the naving convention "name-table" and "name-chart" for divs containing tables and charts.
}
